Fix casing in the 'M.B.Evaluation.GlobResult' class
The correct casing is 'MSBuild' for the 'MSBuildBlob' property.

This is separated from previous casing patches, since, it changes a public property.
It might break VS IDE and others who are depending on the above mentioned property.

The casing change may be cosmetic but it is now easy on the eyes.
There should be an automated way in the editor to force certain words and phrases in preferred casing.