Fix casing in the 'M.B.E.ProjectImportPathMatch' class
The correct casing is 'MSBuild' for the 'MSBuildPropertyFormat' property.

This is separated from previous casing patches, since, the changes are in sources.
The property mentioned above is only used internally. So, it could be safe to fix it.

The casing change may be cosmetic but it is now easy on the eyes.
There should be an automated way in the editor to force certain words and phrases in preferred casing.