Append Project Name to build and publish
When pointing build and publish outside of project directory, we need to check the same
and append the project name to them so that the projects' outputs doesn't clash.

There are no property functions available to check if a path is a parent of another path.
So, we resort to using this hack until a suitable alternative is made available.