Move output paths' trailing slash and mismatch check its own target
We don't need to split them but this will become crowded as we add
output path features later. So, it is best to split them now.

The order of checking the OutputPath properties is now reversed,
meaning the early paths come first instead of the final path properties.

This leads to fail early for the path properties in the order of importance.