Cache Embedded Resources during compilation task
Caching input metadata for `CoreCompile` target is achieved by adding the items
to the `CoreCompileCache` item in the `_GenerateCompileDependencyCache` target.

To cache the `EmbeddedResource` items, we just add those to the existing inputs
through `CoreCompileCache` item, to include them in the cache file which gets
included in the `CoreCompile` target.